+++ app/controllers/ems_block_storage_controller.rb
@@ -2,7 +2,7 @@ class EmsBlockStorageController < ApplicationController
   include Mixins::GenericListMixin
   include Mixins::GenericShowMixin
   include Mixins::EmsCommon
-  include Mixins::EmsCommonAngular
+  include Mixins::EmsCommon::Angular
   include Mixins::GenericSessionMixin
 
   before_action :check_privileges
