+++ app/controllers/ems_container_controller.rb
@@ -2,7 +2,7 @@ class EmsContainerController < ApplicationController
   include Mixins::GenericListMixin
   include Mixins::GenericShowMixin
   include Mixins::EmsCommon # common methods for EmsInfra/Cloud/Container controllers
-  include Mixins::EmsCommonAngular
+  include Mixins::EmsCommon::Angular
   include Mixins::GenericSessionMixin
   include Mixins::DashboardViewMixin
   include ContainersExternalLoggingSupportMixin
