+++ app/controllers/ems_network_controller.rb
@@ -2,7 +2,7 @@ class EmsNetworkController < ApplicationController
   include Mixins::GenericListMixin
   include Mixins::GenericShowMixin
   include Mixins::EmsCommon
-  include Mixins::EmsCommonAngular
+  include Mixins::EmsCommon::Angular
   include Mixins::GenericSessionMixin
 
   before_action :check_privileges
