+++ app/controllers/ems_physical_infra_controller.rb
@@ -2,7 +2,7 @@ class EmsPhysicalInfraController < ApplicationController
   include Mixins::GenericListMixin
   include Mixins::GenericShowMixin
   include Mixins::EmsCommon # common methods for EmsInfra/Cloud controllers
-  include Mixins::EmsCommonAngular
+  include Mixins::EmsCommon::Angular
   include Mixins::GenericSessionMixin
   include Mixins::DashboardViewMixin
